# docs/openapi/trivida-openapi.yaml
openapi: 3.0.3
info:
  title: Trivida API
  version: 1.0.0
  description: >-
    Backend REST API for Trivida MVP (Node/Express + MongoDB). JWT access 15 min, refresh rotation 7 days.
servers:
  - url: https://api.trivida.example.com
    description: Production
  - url: http://localhost:4000
    description: Local Dev
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CronKey:
      type: apiKey
      in: header
      name: X-Cron-Key
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      required: false
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      required: false
    IfMatch:
      name: If-Match
      in: header
      schema: { type: string }
      required: false
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { nullable: true }
...
    FinanceCategory:
      type: object
      required: [_id, userId, name, type]
      properties:
        _id: { type: string }
        userId: { type: string }
        name: { type: string }
        type: { type: string, enum: [expense, income, savings] }
        color: { type: string }
        icon: { type: string }
    Transaction:
      type: object
      required: [_id, userId, categoryId, amountMinor, currency, date]
      properties:
        _id: { type: string }
        userId: { type: string }
        categoryId: { type: string }
        eventId: { type: string }
        amountMinor: { type: integer }
        currency: { type: string }
        date: { type: string, format: date-time }
        note: { type: string }
...
    Event:
      type: object
      required: [_id, userId, title, dateTime, visibility, slug, costPerGuestMinor, currency]
      properties:
        _id: { type: string }
        userId: { type: string }
        title: { type: string }
        dateTime: { type: string, format: date-time }
        timeZone: { type: string }
        location: { type: string }
        visibility: { type: string, enum: [public, private] }
        slug: { type: string }
        costPerGuestMinor: { type: integer }
        currency: { type: string }
        allowPlusOnes: { type: boolean }
        capacity: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventPublic:
      allOf:
        - $ref: '#/components/schemas/Event'
      properties:
        userId: { readOnly: true }
        accessCodeHash: { readOnly: true }
    Invitee:
      type: object
      required: [_id, eventId, name, status, plusOnes]
      properties:
        _id: { type: string }
        eventId: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string, enum: [pending, yes, no, maybe] }
        plusOnes: { type: integer }
...
paths:
...
  /finance/categories:
    get:
      tags: [Finance]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/FinanceCategory' } } } } }
    post:
      tags: [Finance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type: { type: string, enum: [expense, income, savings] }
                color: { type: string }
                icon: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/FinanceCategory' } } } }
  /finance/categories/{id}:
    get:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/FinanceCategory' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                type: { type: string, enum: [expense, income, savings] }
                color: { type: string }
                icon: { type: string }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/FinanceCategory' } } } }
        '412': { description: Precondition failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Deleted }
  /finance/transactions:
    get:
      tags: [Finance]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/Transaction' } }, nextCursor: { type: string } } } } } }
    post:
      tags: [Finance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Transaction'
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
  /finance/transactions/{id}:
    get:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    patch:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Transaction' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
        '412': { description: Precondition failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [Finance]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Deleted }
  /finance/summary:
    get:
      tags: [Finance]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: baseCurrency
          in: query
          schema: { type: string, default: XAF }
        - name: type
          in: query
          schema: { type: string, enum: [expense, income, savings] }
        - name: period
          in: query
          schema: { type: string, enum: [weekly, monthly] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  baseCurrency: { type: string }
                  totalsByType: { type: object, additionalProperties: { type: integer } }
                  totalsByCategory: { type: array, items: { type: object, properties: { categoryId: { type: string }, amountMinor: { type: integer } } } }
                  totalsByCategoryFiltered: { type: array, items: { type: object, properties: { categoryId: { type: string }, amountMinor: { type: integer } } } }
                  timeline:
                    type: array
                    items:
                      type: object
                      properties:
                        periodStart: { type: string, format: date }
                        totalsByType: { type: object, properties: { expense: { type: integer }, income: { type: integer }, savings: { type: integer } } }
  /finance/export.csv:
    get:
      tags: [Finance]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: baseCurrency
          in: query
          schema: { type: string, default: XAF }
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
  /public/events/{slug}:
    get:
      security: []
      tags: [Public]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
        - name: accessCode
          in: query
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/EventPublic' } } } }
        '403': { description: Forbidden (private without code), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /public/events/{slug}/rsvp:
    post:
      security: []
      tags: [Public]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                status: { type: string, enum: [yes, no, maybe], default: yes }
                plusOnes: { type: integer, default: 0 }
                accessCode: { type: string }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Invitee' } } } }
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Invitee' } } } }
        '403': { description: Forbidden (private without/invalid accessCode), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...